First release of River Ruby bindings
A first push of Ruby bindings for River. Meant to be used in conjunction
with a driver like `riverqueue-sequel` [1] to provide an insert-only
client for River. See the README for details on usage.

Overall, I'm happy at how close I was able to keep the API to the Go
version. A lot of syntax in Go just isn't needed due to the more dynamic
and implicit nature of Ruby, but the parts that came through are quite
close. e.g. We have a job args concept, along with `InsertOpts` that can
be added to both jobs and at insert time, just like Go.

Purposely not implemented on this first push (I'll follow up with these
later on):

* Unique jobs.
* Batch insert.

[1] https://github.com/riverqueue/riverqueue-ruby-sequel