Small script to update the version in gemspecs on release (#4)
Add a small script to update the version in gemspecs on release so that
the release instructions don't have to involved manually opening up a
file.

The release on this one is complicated enough between the multiple gems
that it should probably get a full script, but that'll be its own rabbit
hole in testing. This should hold us over.