Implement `#insert_many` for batch job insertion (#5)
Implement `#insert_many` to support inserting jobs in bulk, similar to
what the Go client supports. For better ergonomics, takes either an
array of job args, or an array of `InsertManyParams`, which is a tuple
of job args and insert opts (again, very similar to the Go API).

Also, I seem to have found a way to reimplement the ActiveRecord driver
so that it doesn't have to drop down to Arel, which ends up simplifying
the code quite a bit and making it more alike the Sequel driver.

Lastly, tighten up documentation somewhat by including code example for
the important API methods.