Update development instructions again + update `Gemfile.lock`s
Update development instructions (again) so that we build the gem as part
of the first step so that `Gemfile.lock`s are updated with the new
version and can be checked in as part of the version release.

Also update a few `Gemfile.lock` updates that were left out of the 0.2.0
release due to this omission.