Improve tests for advisory locks (#12)
Improve tests for advisory locks so that we actually check that the
advisory lock is held from another thread. Both ActiveRecord and Sequel
have a lot of laziness, so where an SQL command doesn't have a checked
result, it's easy to think you've invoked it, but only have a lazy data
set instead.