Fix usage of singular `driver/` in CI jobs
One more similar to [1] (previous commit) that fixes the use of the
singular `driver/` in the GitHub workflows YAML file. Also take the
opportunity to rename `ci.yml` to `ci.yaml` so it has a longer quick
find target (one can type "yaml") and to match convention of Go and
Python.

[1] 9154108c937976d11fe155f8a562e00873f0cef2