Test against all supported versions of Ruby
Flesh out the matrix of versions of Ruby that we test the gem against,
adding all three versions that are currently considered to be in
maintenance (3.1, 3.2, 3.3) [1].

The setup-ruby step requires a version be specified, so we have a
slightly awkward situation here were non-spec jobs use a default
`RUBY_VERSION` from env, while the spec job gets a matrix containing
more than one version. I've added comments to `RUBY_VERSION` and the
matrix to hopefully suggest to anyone making a modification that
changing one of them probably means you want to update the other too.

[1] https://www.ruby-lang.org/en/downloads/branches/