Add `make install` target so it's easy to install deps in all directories
Add `make install` target so it's easy to install dependencies in the
main gem and all sub-gems. Change `make bundle-update` to just be `make
update` to fit with the new command, and to make the update version
easier to run (I can't think of anything else it'd potentially clash
with in the future, so I think it's okay to claim the shorter name).